Rename js-compiler plugins to transforms
