+++ packages/@romejs/js-compiler/api/lint.ts
@@ -8,7 +8,7 @@
 import {Program} from '@romejs/js-ast';
 import {PartialDiagnostics} from '@romejs/diagnostics';
 import {TransformRequest} from '../types';
-import {lintTransforms} from '../plugins/lint/index';
+import {lintTransforms} from '../transforms/lint/index';
 import {program} from '@romejs/js-ast';
 import {Cache, Context} from '@romejs/js-compiler';
 import generate from '@romejs/js-generator';
