+++ packages/@romejs/js-compiler/transforms/compile/transpile/arrowFunctions.ts
