+++ packages/@romejs/js-compiler/transforms/compile/transpile/classProperties.ts
