+++ packages/@romejs/js-compiler/transforms/compile/transpile/objectSpread.ts
