+++ packages/@romejs/js-compiler/transforms/compile/transpile/paramlessCatch.ts
