+++ packages/@romejs/js-compiler/transforms/compile/transpile/propertyLiterals.ts
