+++ packages/@romejs/js-compiler/transforms/compileForBundle/assetTransform.ts
