+++ packages/@romejs/js-compiler/transforms/compileForBundle/legacy/inlineRequiresTransform.ts
