+++ packages/@romejs/js-compiler/transforms/compileForBundle/legacy/magicCJSTransform.ts
