+++ packages/@romejs/js-compiler/transforms/compileForBundle/metaPropertyTransform.ts
