+++ packages/@romejs/js-compiler/transforms/compileForBundle/modern/cjsRootTransform.ts
