+++ packages/@romejs/js-compiler/transforms/compileForBundle/modern/esToRefTransform.ts
