+++ packages/@romejs/js-compiler/transforms/compileForBundle/modern/requireRewriteTransform.ts
