+++ packages/@romejs/js-compiler/transforms/compileForBundle/scopedRomeTransform.ts
