+++ packages/@romejs/js-compiler/transforms/defaultHooks/index.ts
