+++ packages/@romejs/js-compiler/transforms/lint/undeclaredVariables.ts
